ContainerStatus: -added STATUS_ABORT PluginsC: -updated initContainer to verify ContainerStatus -updated plugins to make use of STATUS_ABORT
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@35266 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: b565fd5c46d9389387b620916194f6aaa84a4ea9

+++ src/jd/plugins/PluginsC.java
@@ -159,8 +159,8 @@ public abstract class PluginsC {
                 k = key;
             }
             try {
-                callDecryption(file);
-                if (isDeleteContainer(source, file)) {
+                final ContainerStatus cs = callDecryption(file);
+                if (cs != null && cs.isStatus(ContainerStatus.STATUS_FINISHED) && isDeleteContainer(source, file)) {
                     deleteContainer(source, file);
                 }
             } catch (Throwable e) {
